backend/src/requirements.txt
Django==2.2.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2018.5
sqlparse==0.3.0
xlrd==1.2.0
